Fix race condition between variant load and filter tab
In large jobs, the jobTooBig code was being executed before the
filter tab had loaded. The jobTooBig code was trying to set the
text of a div that didn't exist yet, and was failing. Now, the filter
tab is created with once when the page first loads, and again
when the information to fill it with content is available from the
server.